Split product-builder.js out
This component was previously nested inside test-runs-query-builder.js,
which by itself was already not great.

In addition, the test hence had two fixtures and was quite slow compared
to others. Firefox would often hang when running this test on GitHub
Actions.
